Add concurrency options and handle retries
This adds new functionality for controlling the overall concurrency and bumps the default concurrency to a more sane value (1 -> 20). This also fixes an error where some manifests would fail to delete because they reference another image later. This fixes the error by introducing a retry mechanism for manifests that fail to delete with a dangling parent error (3 attempts).